% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGQ_batch.R
\name{EGQ_batch}
\alias{EGQ_batch}
\title{EGQ Multiple Stimuli}
\usage{
EGQ_batch(
  labels,
  urls,
  type = tools::file_ext(url),
  dict = groovescale::groovescale_dict,
  random_order = FALSE,
  show_controls = TRUE,
  allow_download = FALSE,
  ...
)
}
\arguments{
\item{labels}{vector of the labels of the stimuli pages. Duplicates are not
allowed.}

\item{urls}{vector of the URLs of the stimuli. May contain duplicates for
test-retest purposes.}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{random_order}{(Boolean scalar) Whether the order of the items should
be randomized.}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}
}
\description{
This function produces the Experience of Groove Questionnaire
\insertCite{@EGQ; @EGQPaper}{groovescale} for multiple stimuli.
}
\references{
\insertAllCited{}
}
