% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGQ_standalone.R
\name{EGQ_standalone}
\alias{EGQ_standalone}
\title{Standalone Experience of Groove Questionnaire}
\usage{
EGQ_standalone(
  label = "jingle",
 
    url = "https://raw.githubusercontent.com/KilianSander/groovescale/master/inst/www/audio/jingle.mp3",
  type = tools::file_ext(url),
  dict = groovescale::groovescale_dict,
  random_order = FALSE,
  show_controls = TRUE,
  allow_download = FALSE,
  admin_password = "groove",
  researcher_email = NULL,
  languages = groovescale::languages(),
  validate_id = "auto",
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{url}{URL to the audio. Can be an absolute URL (e.g.
"http://mysite.com/audio.mp3") or a URL relative to the /www directory (e.g.
"audio.mp3").}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{random_order}{(Boolean scalar) Whether the order of the items should
be randomized.}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}
}
\description{
This function launches a standalone testing session for the Experience of
Groove Questionnaire \insertCite{EGQPaper}{groovescale}.
}
\references{
\insertAllCited{}
}
