% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groovescale.R
\docType{package}
\name{groovescale}
\alias{groovescale}
\title{Grove Scale Package}
\description{
Contains the Experience of Groove Questionnaire as various psychTestR functions.
}
\author{
Kilian Sander \email{sanderk@stud.hmtm-hannover.de}
}
