% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRV_single_item_pages.R
\name{GRV}
\alias{GRV}
\title{Experience of Groove Questionnaire}
\usage{
GRV(
  url = "https://www.dropbox.com/s/7toxntwv116t6ww/jingle.mp3?dl=0",
  label,
  type = tools::file_ext(url),
  save_answer = TRUE,
  on_complete = NULL,
  arrange_choices_vertically = length(choices) > 2L,
  wait = FALSE,
  loop = FALSE,
  admin_ui = NULL,
  show_controls = TRUE,
  allow_download = FALSE
)
}
\arguments{
\item{url}{URL to the audio. Can be an absolute URL (e.g. "http://mysite.com/audio.mp3") or a URL relative to the /www directory (e.g. "audio.mp3").}

\item{label}{Label to identify all answers corresponding to the audio given under \code{url}}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{save_answer}{Whether or not to save the answer.}

\item{show_controls}{Whether or not to show audio controls to the participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download the audio file; only relevant if \code{show_controls} is \code{TRUE}.}
}
\description{
This functions defines an Experience of Groove Questionnaire module with one page per item for incorporation in a psychTestR timeline.
}
