% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRV_standalone.R
\name{GRV_standalone}
\alias{GRV_standalone}
\title{Standalone Experience of Groove Questionnaire}
\usage{
GRV_standalone(
 
    url = "https://raw.githubusercontent.com/KilianSander/groovescale/master/inst/www/audio/jingle.mp3",
  label = "GRV",
  type = tools::file_ext(url),
  save_answer = TRUE,
  on_complete = NULL,
  arrange_choices_vertically = TRUE,
  wait = FALSE,
  loop = FALSE,
  admin_ui = NULL,
  show_controls = TRUE,
  allow_download = FALSE,
  information = TRUE,
  admin_password = "groove",
  researcher_email = NULL,
  validate_id = "auto",
  languages = groovescale::languages(),
  ...
)
}
\arguments{
\item{url}{URL to the audio. Can be an absolute URL (e.g.
"http://mysite.com/audio.mp3") or a URL relative to the /www directory (e.g.
"audio.mp3").}

\item{label}{Label to identify all answers corresponding to the audio given
under \code{url}}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{save_answer}{Whether or not to save the answer.}

\item{arrange_choices_vertically}{Whether to arrange the response buttons
vertically (the default) as opposed to horizontally.}

\item{wait}{Whether to wait for the audio to finish before displaying the
response buttons.}

\item{loop}{Whether the audio should loop.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}

\item{admin_password}{Password to access the admin panel.}

\item{researcher_email}{Researcher's email; used in participant help message.}

\item{validate_id}{Optional validation function. See \code{validate} in
\link[psychTestR]{page} for further information.}

\item{languages}{Character vector of languages that may be selected via the
URL parameter 'language'. If no language is provided by the URL parameter,
defaults to the first language in this vector. Languages should be encoded
according to ISO 639-2 conventions. Possible languages are shown with
\link[groovescale]{languages}.}

\item{...}{Further arguments to be passed to \code{GRV_standalone()}}

\item{info_page}{If set \code{TRUE} displays a page before the actual
questionnaire informing the participant that the following pages refer
to the same audio example}
}
\description{
This function launches a standalone testing session for the Experience of
Groove Questionnaire.
}
