% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test function}
\usage{
test(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
Character value
}
\description{
More description
}
\examples{
x <- rnorm(n = 1, mean = 0, sd = 1)
test(x)

}
